Use current state value.
After a move from React 0.14 to React 15, the state that was set was not taken from the current state (i.e. setState did not take effect yet).
This change makes it more explicit.
It is possible to convert this component to be pure, and have no state, but I held off on this change as to not break any one's expectations of this component.